Update OrganicChemistryUnknowns.py
Jake Hamilton began to edit the code so that the sql database could be queried through python, but was unable to finish once part of the way into it because Dr. Malmberg had a meeting to attend. I added a description of what was attempted and what errors are likely in rows 27-31